+++ ecommerce-backend/src/main/java/com/comejia/ecommerce/models/entities/Product.java
@@ -1,4 +1,4 @@
-package com.comejia.ecommerce.entities;
+package com.comejia.ecommerce.models.entities;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
