--- ecommerce-backend/src/main/java/com/comejia/ecommerce/services/Service.java
@@ -1,21 +1,0 @@
-package com.comejia.ecommerce.services;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface Service<T> {
-
-    Optional<T> findById(Long id);
-
-    List<T> findAll();
-
-    Optional<T> findByName(String name);
-
-    T save(T entity);
-
-    void delete(T entity);
-
-    void deleteById(Long id);
-
-    void update(T entity, T newEntity);
-}
